+++ packages/react-checkbox/src/components/Checkbox/Checkbox.base.tsx
@@ -1,175 +1,67 @@
 import * as React from 'react';
-import { ICheckboxProps, ICheckboxStyleProps, ICheckboxStyles } from './Checkbox.types';
-import { useControllableValue, useId, useMergedRefs, useWarnings } from '@fluentui/react-hooks';
-import { useFocusRects, classNamesFunction } from '@fluentui/utilities';
-import { Icon } from '@fluentui/react-internal/lib/Icon';
+import { compose, mergeProps } from '@fluentui/react-compose';
+import { ICheckboxProps, ICheckboxSlots, ICheckboxState, ICheckboxSlotProps } from './Checkbox.types';
+import { useCheckbox } from './useCheckbox';
 
-const getClassNames = classNamesFunction<ICheckboxStyleProps, ICheckboxStyles>();
+const defaultSlots: Omit<ICheckboxSlots, 'root'> = {
+  input: 'input',
+  checkmark: 'div',
+  container: 'label',
+  label: 'span',
+  checkbox: 'div',
+};
 
-export const CheckboxBase: React.FunctionComponent<ICheckboxProps> = React.forwardRef<HTMLDivElement, ICheckboxProps>(
-  (props, forwardedRef) => {
-    const {
-      disabled,
-      inputProps,
-      name,
-      ariaLabel,
-      ariaLabelledBy,
-      ariaDescribedBy,
-      ariaPositionInSet,
-      ariaSetSize,
-      title,
-      label,
-      checkmarkIconProps,
-      styles,
-      theme,
-      className,
-      boxSide = 'start',
-    } = props;
-
-    const id = useId('checkbox-', props.id);
-
-    const rootRef = React.useRef<HTMLDivElement | null>(null);
-    const mergedRootRefs: React.Ref<HTMLDivElement> = useMergedRefs(rootRef, forwardedRef);
-    const inputRef = React.useRef<HTMLInputElement>(null);
-
-    const [isChecked, setIsChecked] = useControllableValue(props.checked, props.defaultChecked, props.onChange);
-    const [isIndeterminate, setIsIndeterminate] = useControllableValue(props.indeterminate, props.defaultIndeterminate);
-
-    useFocusRects(rootRef);
-    useDebugWarning(props);
-    useComponentRef(props, isChecked, isIndeterminate, inputRef);
-
-    const classNames = getClassNames(styles!, {
-      theme: theme!,
-      className,
-      disabled,
-      indeterminate: isIndeterminate,
-      checked: isChecked,
-      reversed: boxSide !== 'start',
-      isUsingCustomLabelRender: !!props.onRenderLabel,
-    });
-
-    const onChange = (ev: React.ChangeEvent<HTMLElement>): void => {
-      if (isIndeterminate) {
-        // If indeterminate, clicking the checkbox *only* removes the indeterminate state (or if
-        // controlled, lets the consumer know to change it by calling onChange). It doesn't
-        // change the checked state.
-        setIsChecked(!!isChecked, ev);
-        setIsIndeterminate(false);
-      } else {
-        setIsChecked(!isChecked, ev);
-      }
-    };
-
-    const defaultLabelRenderer = React.useCallback(
-      (checkboxProps?: ICheckboxProps): JSX.Element | null => {
-        if (!checkboxProps) {
-          return null;
-        }
-        return checkboxProps.label ? (
-          <span aria-hidden="true" className={classNames.text} title={checkboxProps.title}>
-            {checkboxProps.label}
-          </span>
-        ) : null;
-      },
-      [classNames.text],
+export const CheckboxBase = compose<'div', ICheckboxProps, {}, ICheckboxProps, {}>(
+  (props, forwardedRef, composeOptions) => {
+    const { slotProps, slots } = mergeProps<ICheckboxProps, ICheckboxState, ICheckboxSlots, ICheckboxSlotProps>(
+      composeOptions.state,
+      composeOptions,
     );
 
-    const onRenderLabel = props.onRenderLabel || defaultLabelRenderer;
-
-    const ariaChecked: React.InputHTMLAttributes<HTMLInputElement>['aria-checked'] = isIndeterminate
-      ? 'mixed'
-      : isChecked
-      ? 'true'
-      : 'false';
-
-    const slotProps = {
-      root: {
-        className: classNames.root,
-        title,
-        ref: mergedRootRefs,
-      },
-      input: {
-        className: classNames.input,
-        type: 'checkbox' as React.InputHTMLAttributes<HTMLInputElement>['type'],
-        ...inputProps,
-        ref: inputRef,
-        checked: !!isChecked,
-        disabled,
-        name,
-        id,
-        title,
-        onChange,
-        'data-ktp-execute-target': true,
-        'aria-disabled': disabled,
-        'aria-label': ariaLabel || label,
-        'aria-labelledby': ariaLabelledBy,
-        'aria-describedby': ariaDescribedBy,
-        'aria-posinset': ariaPositionInSet,
-        'aria-setsize': ariaSetSize,
-        'aria-checked': ariaChecked,
-      },
-      checkbox: {
-        className: classNames.checkbox,
-        'data-ktp-target': true,
-      },
-      container: {
-        className: classNames.label,
-        htmlFor: id,
-      },
+    const onRenderLabel = (): JSX.Element => {
+      return <slots.label {...slotProps.label} />;
     };
 
     return (
-      <div {...slotProps.root}>
-        <input {...slotProps.input} />
-        <label {...slotProps.container}>
-          <div {...slotProps.checkbox}>
-            <Icon iconName="CheckMark" {...checkmarkIconProps} className={classNames.checkmark} />
-          </div>
-          {onRenderLabel(props, defaultLabelRenderer)}
-        </label>
-      </div>
+      <slots.root {...slotProps.root}>
+        <slots.input {...slotProps.input} />
+        <slots.container {...slotProps.container}>
+          <slots.checkbox {...slotProps.checkbox}>
+            <slots.checkmark {...slotProps.checkmark} />
+          </slots.checkbox>
+
+          {// eslint-disable-next-line deprecation/deprecation
+          (props.onRenderLabel || onRenderLabel)(props, onRenderLabel)}
+        </slots.container>
+      </slots.root>
     );
+  },
+  {
+    slots: defaultSlots,
+    state: useCheckbox,
+    displayName: 'CheckboxBase',
+    handledProps: [
+      'componentRef',
+      'onChange',
+      'checked',
+      'defaultChecked',
+      'boxSide',
+      'label',
+      'checkmark',
+      'disabled',
+      'inputProps',
+      'boxSide',
+      'ariaLabel',
+      'ariaLabelledBy',
+      'ariaDescribedBy',
+      'ariaPositionInSet',
+      'ariaSetSize',
+      'styles',
+      'onRenderLabel',
+      'indeterminate',
+      'defaultIndeterminate',
+    ],
   },
 );
 
-CheckboxBase.displayName = 'CheckboxBase';
-
-function useDebugWarning(props: ICheckboxProps) {
-  if (process.env.NODE_ENV !== 'production') {
-    // eslint-disable-next-line react-hooks/rules-of-hooks -- build-time conditional
-    useWarnings({
-      name: 'Checkbox',
-      props,
-      mutuallyExclusive: {
-        checked: 'defaultChecked',
-        indeterminate: 'defaultIndeterminate',
-      },
-    });
-  }
-}
-
-function useComponentRef(
-  props: ICheckboxProps,
-  isChecked: boolean | undefined,
-  isIndeterminate: boolean | undefined,
-  checkBoxRef: React.RefObject<HTMLInputElement>,
-) {
-  React.useImperativeHandle(
-    props.componentRef,
-    () => ({
-      get checked() {
-        return !!isChecked;
-      },
-      get indeterminate() {
-        return !!isIndeterminate;
-      },
-      focus() {
-        if (checkBoxRef.current) {
-          checkBoxRef.current.focus();
-        }
-      },
-    }),
-    [checkBoxRef, isChecked, isIndeterminate],
-  );
-}
+CheckboxBase.defaultProps = {};
