+++ packages/react-checkbox/src/components/Checkbox/Checkbox.scss
