--- packages/react-checkbox/src/next.ts
@@ -1,1 +1,0 @@
-export * from './next/index';
