--- packages/react-checkbox/src/next/Checkbox.tsx
@@ -1,19 +1,0 @@
-import * as React from 'react';
-import { compose, ComposeOptions } from '@fluentui/react-compose';
-import { CheckMarkIcon } from '@fluentui/react-icons-mdl2';
-import { CheckboxBase } from './Checkbox.base';
-import { useCheckboxClasses } from './useCheckboxClasses';
-import { ICheckboxProps } from './Checkbox.types';
-
-const composeOptions: ComposeOptions = {
-  displayName: 'Checkbox',
-  classes: useCheckboxClasses,
-  slots: {
-    checkmark: CheckMarkIcon,
-  },
-};
-
-export const Checkbox: React.FunctionComponent<ICheckboxProps> = compose<'div', ICheckboxProps, {}, ICheckboxProps, {}>(
-  CheckboxBase,
-  composeOptions,
-);
