+++ packages/react-examples/src/react/Checkbox/Checkbox.Indeterminate.Example.tsx
