+++ packages/react-internal/src/components/Checkbox/Checkbox.tsx
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import { styled } from '@fluentui/utilities';
+import { CheckboxBase } from './Checkbox.base';
+import { getStyles } from './Checkbox.styles';
+import { ICheckboxProps, ICheckboxStyleProps, ICheckboxStyles } from './Checkbox.types';
+
+export const Checkbox: React.FunctionComponent<ICheckboxProps> = styled<
+  ICheckboxProps,
+  ICheckboxStyleProps,
+  ICheckboxStyles
+>(CheckboxBase, getStyles, undefined, { scope: 'Checkbox' });
